prerender initial data in infinite scroll
